server/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.38.0
langchain-core==1.0.0
pydantic==2.12.3
python-dotenv==1.1.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pymysql==1.1.1
cryptography==44.0.0
openai>=1.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.0
email-validator>=2.0.0
bcrypt===4.0.1
colorama>=0.4.6
